requirements.txt
pandas
python-dotenv
langchain-google-genai
pdfplumber
duckdb